fix(logging): Only log stack traces for server errors (#8)
We've noticed that go projects configured for Sentry log the stack
trace of every error returned by a handler. This includes "errors"
that are completely normal, such as 401s, 404s, etc. This is the
result of calling `.Err()` on the logger, which captures the stack.

This commit changes this behavior to only call `.Err()` on an
actual server error; that is, a 5xx status or a non-HTTP error.