feat(responses): Only write JSON response if not previously sent
When the Sentry handler is registered as the Echo ErrorHandler, it
is the last stop when a handler function returns an error. We were
previously relying on it to _always_ send a JSON response, as well.
This works for basic cases, but is a bit overly prescriptive.

This commit changes the output behavior in two ways.

First, the response is only written by the middleware if the
response was not already written. This allows handlers to capture
an error, write an appropriate response (error page, JSON, etc),
and _still_ report it to Sentry.

Second, if the Error implements the `json.Marshaller` interface,
it is marshalled as the response directly. This allows users of the
sentry-echo package much greater control over how API errors are
formatted.

Additionally, if the error is an HTTPError, we unwrap it's
`Internal` error, if present. This ensures that the underlying
error is the one actually reported to Sentry.